feat(quay): Refractor some code
